requirement.txt
amqp==2.4.2
backports.csv==1.0.7
billiard==3.6.0.0
celery==4.1.1
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111

django-crispy-forms==1.7.2
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.2.0
django-redis==4.10.0
django-redis-sessions==0.6.1
django-tinymce==2.8.0
et-xmlfile==1.0.1
fdfs-client-py==1.2.6
future==0.17.1
httplib2==0.12.3
idna==2.8
itsdangerous==1.1.0
jdcal==1.4.1
jieba==0.39
kombu==4.5.0
Markdown==3.1.1
mutagen==1.42.0
mysql-client==0.0.1
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.0.0
PyMySQL==0.9.3
pytz==2019.1
PyYAML==5.1
redis==3.2.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
urllib3==1.25.3
uWSGI==2.0.18
vine==1.3.0
Whoosh==2.7.4
xlrd==1.2.0
xlwt==1.3.0
